Only remove and symlink if symbolic link doesn't exist in ui Change purple to magenta in ripgrep Add success feedback on deployment

+++ .config/fish/entgra.fish
@@ -70,11 +70,10 @@ function emi
 	end
 	cd $curr_dir
 end
-# build given packages by path (i.e emif HEAD $carbon <path-to-package>)
+# build given packages by path (i.e emif <path-to-package>)
 function emif
 	set -l curr_dir (pwd)
-	cd $argv[2]
-	for x in $argv[3..-1]
+	for x in $argv
 		echo "[BUILDING] $x" | rg "BUILDING" --passthru --colors 'match:fg:magenta' --color always
 		cd $x
 		mvn clean install -Dmaven.test.skip=true
@@ -116,6 +115,9 @@ function entup
 					echo "[COPYING] $ui/target/$war --> $warBundles/$ui-$war" | rg "$work" --replace "" | rg "COPYING|-->" --passthru --colors 'match:fg:green' --color always
 					cp $dir/target/$war $warBundles/$ui_name
 				end
+				echo "------------------------------------------------" | rg "-" --colors 'match:fg:white'
+				echo " DEPLOYED SUCCESSFULLY" | rg "DEPLOYED SUCCESSFULLY" --passthru --colors 'match:fg:magenta' --color always | rg "INFO" --passthru --colors 'match:fg:blue' --color always
+				echo "------------------------------------------------" | rg "-" --colors 'match:fg:white'
 			else if test "$war" != ""
 				set -l distWar (ls $warBundles/ | rg "^$war")
 				set -l warDir (echo $war | rg '.war' --replace '')
@@ -134,6 +136,9 @@ function entup
 				end
 				echo "[COPYING] $dir/target/$war --> $warBundles" | rg "$work" --replace "" | rg "COPYING|-->" --passthru --colors 'match:fg:green' --color always
 				cp $dir/target/$war $warBundles
+				echo "------------------------------------------------" | rg "-" --colors 'match:fg:white'
+				echo " DEPLOYED SUCCESSFULLY" | rg "DEPLOYED SUCCESSFULLY" --passthru --colors 'match:fg:magenta' --color always | rg "INFO" --passthru --colors 'match:fg:blue' --color always
+				echo "------------------------------------------------" | rg "-" --colors 'match:fg:white'
 			else
 				set -l jar (ls $dir/target | rg '.*jar') 
 				if test "$jar" != ""
@@ -150,6 +155,9 @@ function entup
 					mkdir $patches/$patchDir
 					echo "[COPYING] $dir/target/$jar --> $patches/$patchDir" | rg "$work" --replace "" | rg "COPYING|-->" --passthru --colors 'match:fg:green' --color always
 					cp $dir/target/$jar $patches/$patchDir
+					echo "------------------------------------------------" | rg "-" --colors 'match:fg:white'
+					echo " DEPLOYED SUCCESSFULLY" | rg "DEPLOYED SUCCESSFULLY" --passthru --colors 'match:fg:magenta' --color always | rg "INFO" --passthru --colors 'match:fg:blue' --color always
+					echo "------------------------------------------------" | rg "-" --colors 'match:fg:white'
 				else
 					echo "[ERROR] No target found" | rg "ERROR" --passthru --colors 'match:fg:255,51,71' --color always
 					return 1
@@ -165,8 +173,7 @@ end
 # deploy selected packages by path
 function entupf
 	set -l curr_dir (pwd)
-	cd $argv[2]
-	for x in $argv[3..-1]
+	for x in $argv
 		echo "[UPDATING] $x" | rg "$work" --replace "" | rg "UPDATING" --passthru --colors 'match:fg:0,229,255' --color always
 		set -l war (ls $x/target | rg '.*war') 
 		# ui-request-handler deployment exception
@@ -183,7 +190,6 @@ function entupf
 			end
 			if test "$dirWarRm" != ""
 				echo "[DELETING] $dirWarRm" | rg "$work" --replace "" | rg "DELETING" --passthru --colors 'match:fg:255,51,71' --color always
-
 				rm -rf $dirWarRm
 			end
 			for ui in $UI
@@ -191,6 +197,9 @@ function entupf
 				echo "[COPYING] $ui/target/$war --> $warBundles/$ui-$war" | rg "$work" --replace "" | rg "COPYING|-->" --passthru --colors 'match:fg:green' --color always
 				cp $x/target/$war $warBundles/$ui_name
 			end
+			echo "------------------------------------------------" | rg "-" --colors 'match:fg:white'
+			echo " DEPLOYED SUCCESSFULLY" | rg "DEPLOYED SUCCESSFULLY" --passthru --colors 'match:fg:magenta' --color always | rg "INFO" --passthru --colors 'match:fg:blue' --color always
+			echo "------------------------------------------------" | rg "-" --colors 'match:fg:white'
 		else if test "$war" != ""
 			set -l distWar (ls $warBundles/ | rg "^$war")
 			set -l warDir (echo $war | rg '.war' --replace '')
@@ -209,6 +218,9 @@ function entupf
 			end
 			echo "[COPYING] $x/target/$war --> $warBundles" | rg "$work" --replace "" | rg "COPYING|-->" --passthru --colors 'match:fg:green' --color always
 			cp $x/target/$war $warBundles
+			echo "------------------------------------------------" | rg "-" --colors 'match:fg:white'
+			echo " DEPLOYED SUCCESSFULLY" | rg "DEPLOYED SUCCESSFULLY" --passthru --colors 'match:fg:magenta' --color always | rg "INFO" --passthru --colors 'match:fg:blue' --color always
+			echo "------------------------------------------------" | rg "-" --colors 'match:fg:white'
 		else
 			set -l jar (ls $x/target | rg '.*jar') 
 			if test "$jar" != ""
@@ -225,6 +237,9 @@ function entupf
 				mkdir $patches/$patchDir
 				echo "[COPYING] $x/target/$jar --> $patches/$patchDir" | rg "$work" --replace "" | rg "COPYING|-->" --passthru --colors 'match:fg:green' --color always
 				cp $x/target/$jar $patches/$patchDir
+				echo "------------------------------------------------" | rg "-" --colors 'match:fg:white'
+				echo " DEPLOYED SUCCESSFULLY" | rg "DEPLOYED SUCCESSFULLY" --passthru --colors 'match:fg:magenta' --color always | rg "INFO" --passthru --colors 'match:fg:blue' --color always
+				echo "------------------------------------------------" | rg "-" --colors 'match:fg:white'
 			else
 				echo "[ERROR] No target found" | rg "ERROR" --passthru --colors 'match:fg:255,51,71' --color always
 				return 1
@@ -251,11 +266,11 @@ function ebdf
 	set -l curr_dir (pwd)
 	set dirs (etb $argv | fzf -m --reverse)
 	for x in $dirs
-		emif $argv $x
+		emif $x
 		if test "$status" -gt 0
 			return 1
 		end
-		entupf $argv $x
+		entupf $x
 		if test "$status" -gt 0
 			return 1
 		end
@@ -267,16 +282,24 @@ function ebdfa
 	set -l curr_dir (pwd)
 	set dirs (etba $argv | fzf -m --reverse)
 	for x in $dirs
-		emif $argv $x
+		emif $x
 		if test "$status" -gt 0
 			return 1
 		end
-		entupf $argv $x
+		entupf $x
 		if test "$status" -gt 0
 			return 1
 		end
 	end
 	cd $curr_dir
+end
+
+function entapply
+	if test "$argv[1]" = "$carbon" || test "$argv[1]" = "$plugin" || test "$argv[1]" = "$emm"
+		ebdfa $argv
+	else
+		ebdf $argv
+	end
 end
 
 # entgra ui watch
@@ -291,7 +314,7 @@ function entuiwatch
 
 		if test -z "$builtIndex"
 			echo "[INFO] $ui/react-app/dist is empty" | rg "INFO" --colors match:fg:blue
-			echo "[BUILDING] npm run dev on $warDir" | rg "BUILDING" --colors match:fg:purple
+			echo "[BUILDING] npm run dev on $warDir" | rg "BUILDING" --colors match:fg:magenta
 			npm run --prefix $emm/components/ui/$ui/react-app dev
 		end
 		
@@ -299,16 +322,19 @@ function entuiwatch
 		$warBundles/$warDir/main.js $warBundles/$warDir/main.css.map \
 		$warBundles/$warDir/main.js.map
 
+		if ! test -L $reactApp/index.html && ! test -L $reactApp/main.css && ! test -L $reactApp/main.js &&
+			! test -L $reactApp/main.css.map && ! test -L $reactApp/main.js.map
 
-		echo "[REMOVING] $rmWarDirFiles" | rg "REMOVING" --colors match:fg:red
-		rm -rf $rmWarDirFiles
+			echo "[REMOVING] $rmWarDirFiles" | rg "REMOVING" --colors match:fg:red
+			rm -rf $rmWarDirFiles
 
-		echo "[LINKING] Linking $ui/react-app/dist --> $warDir" | rg "LINKING|-->" --colors match:fg:blue
-		ln -fs $warBundles/$warDir/index.html   $reactApp/index.html 	
-		ln -fs $warBundles/$warDir/main.css     $reactApp/main.css 		
-		ln -fs $warBundles/$warDir/main.js      $reactApp/main.js 		
-		ln -fs $warBundles/$warDir/main.css.map $reactApp/main.css.map 	
-		ln -fs $warBundles/$warDir/main.js.map  $reactApp/main.js.map 	
+			echo "[LINKING] Linking $ui/react-app/dist --> $warDir" | rg "LINKING|-->" --colors match:fg:blue
+			ln -fs $warBundles/$warDir/index.html   $reactApp/index.html 	
+			ln -fs $warBundles/$warDir/main.css     $reactApp/main.css 		
+			ln -fs $warBundles/$warDir/main.js      $reactApp/main.js 		
+			ln -fs $warBundles/$warDir/main.css.map $reactApp/main.css.map 	
+			ln -fs $warBundles/$warDir/main.js.map  $reactApp/main.js.map 	
+		end
 
 		echo "[WATCHING] npm run watch on $warDir" | rg "WATCHING" --colors match:fg:green
 		npm run --prefix $emm/components/ui/$ui/react-app watch
